fix(parser): share XML escape helpers
