+++ internal/i18n/translationfileparser/generic_xml_parser.go
@@ -514,14 +514,3 @@ func isGenericXMLSpecializedRoot(name string) bool {
 		return false
 	}
 }
-
-func escapeXMLAttr(s string) string {
-	replacer := strings.NewReplacer(
-		"&", "&amp;",
-		"<", "&lt;",
-		">", "&gt;",
-		`"`, "&quot;",
-		"'", "&apos;",
-	)
-	return replacer.Replace(s)
-}
